Move HTML validation “all component” tests into `components.template.test.js`
These tests don’t need a Chromium browser window (Puppeteer environment)

This change brings Nunjucks HTML render + validation together
